TraderDAOIntTest

+++ springboot/src/main/java/ca/jrvs/apps/trading/model/domain/IexQuote.java
@@ -145,11 +145,11 @@ public class IexQuote {
   @JsonProperty("avgTotalVolume")
   private Long avgTotalVolume;
   @JsonProperty("iexBidPrice")
-  private Long iexBidPrice;
+  private Double iexBidPrice;
   @JsonProperty("iexBidSize")
   private Long iexBidSize;
   @JsonProperty("iexAskPrice")
-  private Long iexAskPrice;
+  private Double iexAskPrice;
   @JsonProperty("iexAskSize")
   private Long iexAskSize;
   @JsonProperty("iexOpen")
@@ -495,11 +495,11 @@ public class IexQuote {
     this.avgTotalVolume = avgTotalVolume;
   }
 
-  public Long getIexBidPrice() {
+  public Double getIexBidPrice() {
     return iexBidPrice;
   }
 
-  public void setIexBidPrice(Long iexBidPrice) {
+  public void setIexBidPrice(Double iexBidPrice) {
     this.iexBidPrice = iexBidPrice;
   }
 
@@ -511,11 +511,11 @@ public class IexQuote {
     this.iexBidSize = iexBidSize;
   }
 
-  public Long getIexAskPrice() {
+  public Double getIexAskPrice() {
     return iexAskPrice;
   }
 
-  public void setIexAskPrice(Long iexAskPrice) {
+  public void setIexAskPrice(Double iexAskPrice) {
     this.iexAskPrice = iexAskPrice;
   }
 
@@ -614,6 +614,5 @@ public class IexQuote {
   public void setUSMarketOpen(Boolean USMarketOpen) {
     isUSMarketOpen = USMarketOpen;
   }
-
 }
